Feature ETP-435: Improved the structure
